use uint8 like defined in messages: fix warning C4244

+++ rclcpp_lifecycle/include/rclcpp_lifecycle/lifecycle_node.hpp
@@ -356,7 +356,7 @@ public:
 
   RCLCPP_LIFECYCLE_PUBLIC
   const State &
-  trigger_transition(unsigned int transition_id);
+  trigger_transition(uint8_t transition_id);
 
   RCLCPP_LIFECYCLE_PUBLIC
   bool
